+++ includes/admin/menu/notifications.php
@@ -8,33 +8,16 @@
     </p>
 
 
-    <table class="awca_notification_list_table">
-        <thead>
-        <tr style="text-align:right">
-            <th>ردیف</th>
-            <th>عنوان</th>
-            <th>متن</th>
-        </tr>
-        </thead>
+    <div class="list" id="awca_notification_list">
 
-        <tbody class="list" id="awca_notification_list">
+        <div class="awca-loading-message-spinner">
+            در حال دریافت اعلان ها ...
+            <svg class="spinner-loading" width="24px" height="24px" viewBox="0 0 66 66" xmlns="http://www.w3.org/2000/svg">
+                <circle class="path" fill="none" stroke-width="6" stroke-linecap="round" cx="33" cy="33" r="30"></circle>
+            </svg>
+        </div>
 
-            <tr class="item">
-                <td></td>
-                <td>
-                    <div class="awca-loading-message-spinner">
-                        در حال دریافت اعلان ها ...
-                        <svg class="spinner-loading" width="24px" height="24px" viewBox="0 0 66 66" xmlns="http://www.w3.org/2000/svg">
-                            <circle class="path" fill="none" stroke-width="6" stroke-linecap="round" cx="33" cy="33" r="30"></circle>
-                        </svg>
-                    </div>
-                </td>
-                <td></td>
-            </tr>
-
-        </tbody>
-
-    </table>
+    </div>
 
     <div id="awca_pagination"></div>